+++ src/ducks/Studio/Tabs/Widgets.module.scss
@@ -1,0 +1,21 @@
+.content {
+  display: flex;
+  flex: 1;
+}
+
+.widgets {
+  flex: 1;
+}
+
+.side {
+  background: var(--white);
+  border-radius: 4px;
+  border: 1px solid var(--porcelain);
+  box-shadow: 0 2px 8px rgba(59, 84, 106, 0.03),
+    0 1px 2px rgba(59, 84, 106, 0.03);
+  margin-left: 16px;
+  width: 355px;
+  position: sticky !important;
+  top: 67px !important;
+  max-height: calc(100vh - 54px - 13px - 13px - 12px);
+}
