+++ src/ducks/Studio/Widget/ChartWidget.js
@@ -1,15 +1,15 @@
 import React, { useMemo, useEffect, useState } from 'react'
 import Widget from './Widget'
 import { newWidget } from './utils'
+import StudioChart from '../Chart'
 import { dispatchWidgetMessage } from '../widgetMessage'
-import { Metric } from '../../dataHub/metrics'
+import { DEFAULT_OPTIONS } from '../defaults'
+import { useTimeseries } from '../timeseries/hooks'
+import { buildAnomalies } from '../timeseries/anomalies'
+import { buildComparedMetric } from '../Compare/utils'
+import { buildChartShareLink } from '../url/generate'
 import { useClosestValueData } from '../../Chart/hooks'
-import StudioChart from '../../Studio/Chart'
-import { useTimeseries } from '../../Studio/timeseries/hooks'
-import { buildAnomalies } from '../../Studio/timeseries/anomalies'
-import { DEFAULT_OPTIONS } from '../../Studio/defaults'
-import { buildComparedMetric } from '../../Studio/Compare/utils'
-import { buildChartShareLink } from '../../Studio/url/generate'
+import { Metric } from '../../dataHub/metrics'
 import { MirroredMetric } from '../../dataHub/metrics/mirrored'
 
 export const Chart = ({
