+++ src/ducks/Studio/Widget/Widget.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import cx from 'classnames'
-import styles from '../index.module.scss'
+import styles from './Widget.module.scss'
 
 const Widget = ({ className, children }) => (
   <div className={cx(styles.widget, className)}>{children}</div>
