+++ src/ducks/Studio/Widget/Widget.module.scss
@@ -1,0 +1,18 @@
+.widget {
+  background: var(--white);
+  padding: 20px;
+  border-radius: 4px;
+  border: 1px solid var(--porcelain);
+  box-shadow: 0 2px 8px rgba(59, 84, 106, 0.03),
+    0 1px 2px rgba(59, 84, 106, 0.03);
+  margin-top: 16px;
+  height: 530px;
+
+  &:only-child {
+    height: 100%;
+  }
+
+  &:first-child {
+    margin: 0;
+  }
+}
