+++ src/ducks/Studio/Widget/utils.js
