+++ src/ducks/Studio/widgetMessage.js
