--- src/ducks/Studio2/index.module.scss
@@ -1,85 +1,0 @@
-@import '~@santiment-network/ui/variables';
-@import '~@santiment-network/ui/mixins';
-
-.wrapper {
-  display: flex;
-  position: relative;
-}
-
-.main {
-  position: relative;
-  min-height: 100vh;
-  flex: 1;
-}
-
-.tab {
-  background: var(--athens);
-  padding: 25px;
-  min-height: 100vh;
-}
-
-.widget {
-  background: var(--white);
-  padding: 20px;
-  border-radius: 4px;
-  border: 1px solid var(--porcelain);
-  box-shadow: 0 2px 8px rgba(59, 84, 106, 0.03),
-    0 1px 2px rgba(59, 84, 106, 0.03);
-  margin-top: 16px;
-  height: 530px;
-
-  &:only-child {
-    height: 100%;
-  }
-
-  &:first-child {
-    margin: 0;
-  }
-}
-
-// explain metrics
-.content {
-  display: flex;
-  flex: 1;
-}
-
-.widgets {
-  flex: 1;
-}
-
-.side {
-  background: var(--white);
-  /* padding: 20px; */
-  border-radius: 4px;
-  border: 1px solid var(--porcelain);
-  box-shadow: 0 2px 8px rgba(59, 84, 106, 0.03),
-    0 1px 2px rgba(59, 84, 106, 0.03);
-  margin-left: 16px;
-  width: 355px;
-  position: sticky !important;
-  top: 67px !important;
-  // height: calc(100vh - 67px - 16px);
-  max-height: calc(100vh - 54px - 13px - 13px - 12px);
-}
-
-// holders
-
-.holders {
-  position: relative;
-
-  &_opened {
-    padding-right: 340px !important;
-  }
-}
-
-.toggle {
-  left: initial !important;
-  right: 0;
-  width: 24px !important;
-}
-
-.title {
-  font-weight: 700;
-  display: flex;
-  align-items: center;
-}
